+++ src/parac/__init__.py
@@ -1,5 +1,5 @@
 # coding=utf-8
-""" Init file for managing both compiler and preprocessor """
+""" Para-C (parac) Module for the Para-C Compiler and Pre-Processor """
 
 __title__ = "parac"
 __description__ = "Para-C Module"
@@ -7,7 +7,7 @@ __url__ = "https://github.com/Para-C/Para-C/"
 __author__ = "Luna Klatzer"
 __author_email__ = "luna.klatzer@gmail.com"
 __license__ = "GNU GENERAL PUBLIC LICENSE v3.0"
-__version__ = "v0.1.dev1"
+__version__ = "v0.1.dev2"
 __code_name__ = ""
 __release__ = f"{__code_name__} {__version__}"
 __copyright__ = "Luna Klatzer"
@@ -55,6 +55,7 @@ __all__ = [
     'DEFAULT_BUILD_PATH',
     'DEFAULT_CONFIG',
     'C_COM_EXISTENCE_OVERWRITE',
+    'RUNTIME_COMPILER',
     *exceptions.__all__,
     *const.__all__,
     *MODULES
@@ -65,3 +66,7 @@ import colorama
 
 colorama.init(autoreset=True)
 lib_logging.getLogger(__name__).addHandler(lib_logging.NullHandler())
+
+# An instance of the compiler, which should be generally used in the module
+# due to the logging logic
+RUNTIME_COMPILER: compiler.ParacCompiler = compiler.ParacCompiler()